Disable crt secure warnings

+++ remus/client/ServerConnection.cxx
@@ -101,7 +101,7 @@ remus::client::ServerConnection make_ServerConnection(const std::string& dest)
 //------------------------------------------------------------------------------
 boost::shared_ptr<zmq::context_t> make_ServerContext(std::size_t threads)
 {
-  return boost::make_shared<zmq::context_t>(threads);
+  return boost::make_shared<zmq::context_t>( static_cast<int>(threads) );
 }
